fix: corrige overflow de inteiro e alocação ilimitada de map no decoder (#3)
Em 27/07/2026 saíram duas advisories contra `github.com/hamba/avro/v2`
como "all versions, no known fixed version": GO-2026-5047 (CVE-2026-46384,
CVSS 8.7) e GO-2026-5048. O upstream está descontinuado e só corrigiu no
fork iskorotkov/avro, então a correção foi portada para cá — trocar o
module path por um fork de terceiro contrariaria a manutenção interna. As
duas advisories não têm sinal automatizado neste repo: depois da renomeação
para arquivei/avro o module path não bate mais com a OSV, e o govulncheck
não reportaria vulnerabilidade do próprio main module de qualquer forma.
Ambas foram achadas lendo as advisories e conferindo o código à mão.

GO-2026-5047 é a mesma falha em seis lugares: o decoder lê um valor de 64
bits do wire, estreita para int antes de validar, e valida o valor
estreitado. Em builds de 32 bits — 386 e arm estão na matriz do CI — o
estreitamento trunca, e (1<<32)+5 vira 5, então o limite é conferido
contra um valor que o decoder nunca usa. Todos os pontos passaram a
validar sobre o int64 e só estreitar depois, o que torna o comportamento
idêntico nas duas plataformas. O caso mais grave em amd64 é readBytes com
MaxByteSliceSize desabilitado, onde um length acima do heap alocável chega
a make e panica com "makeslice: len out of range".

Um dos sub-issues não é DoS, é corrupção silenciosa de dados: no decode
genérico o índice de union era estreitado antes do bounds-check, e 1<<32
virava 0 em 32 bits, selecionando types[0]. Em union nullable
["null", T] — a forma idiomática — o consumidor recebe nil onde o produtor
codificou payload, sem erro nenhum.

GO-2026-5048 é o crescimento de map sem limite superior, que passa a ser
controlado por Config.MaxMapAllocSize, acumulado entre blocos para que não
dê para burlar fatiando as entradas em blocos individualmente pequenos.
O limite é opt-in e o default segue ilimitado: MaxSliceAllocSize, o
análogo para arrays, já usa essa convenção, e qualquer default quebraria
silenciosamente quem decodifica maps grandes e legítimos — é uma contagem
de entradas, sem valor obviamente alto o bastante como os limites em
bytes. Só subir a versão portanto não protege; consumidores de entrada não
confiável precisam setar o limite explicitamente. A severidade real aqui é
menor que a descrita na advisory, porque estes decoders nunca pré-alocaram
a partir do count — o que sobra é amplificação, não alocação instantânea.

A API pública não quebra: só entra Reader.SkipNBytesInt64. ReadBlockHeader
mantém a assinatura, mas passa a reportar erro em headers fora da faixa
int32, que nenhum encoder Avro legítimo produz. Fica uma limitação de
verificação: os testes do pacote raiz não compilam sob GOARCH=386 por três
arquivos pré-existentes que usam constantes que estouram int em 32 bits,
então os sub-issues exclusivos de 32 bits não têm teste de ponta a ponta
naquela arquitetura.

Links:
  + https://pkg.go.dev/vuln/GO-2026-5047
  + https://pkg.go.dev/vuln/GO-2026-5048

+++ reader_generic.go
@@ -97,8 +97,13 @@ func (r *Reader) ReadNext(schema Schema) any {
 		return obj
 	case Union:
 		types := schema.(*UnionSchema).Types()
-		idx := int(r.ReadLong())
-		if idx < 0 || idx > len(types)-1 {
+		// The index is bounds-checked as an int64. Narrowing first would let
+		// an out-of-range wire index truncate to a valid one on 32-bit builds
+		// — 1<<32 becomes 0 — silently selecting types[0], which for the
+		// idiomatic ["null", T] union yields null where the producer encoded a
+		// payload.
+		idx := r.ReadLong()
+		if idx < 0 || idx > int64(len(types)-1) {
 			r.ReportError("Read", "unknown union type")
 			return nil
 		}
